% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureAE.R
\name{AE}
\alias{AE}
\title{Feature auto encoder class}
\usage{
AE(dim)
}
\arguments{
\item{dim}{Dimension of features of input.}
}
\value{
A new \code{AE} object.

encoded matrix

reconstructed matrix
}
\description{
Create a new feature autoencoder object.

Perform demension reduction to encode feature through the first two layers

Reconstruct to decode feature through the last two layers

Forward feeding
}
\details{
An auto-encoder to perform dimensional reeduction on features

#' @references
\insertRef{GCN}{scRGNet}
}
\section{Fields}{

\describe{
\item{\code{dim}}{Dimension of features of input matrix}

\item{\code{fc1}}{First layer of encoder}

\item{\code{fc2}}{Hidden layer decoder}

\item{\code{fc3}}{Hidden layer decoder}

\item{\code{fc4}}{Second layer decoder}
}}

