% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{calculate_knn_graph}
\alias{calculate_knn_graph}
\title{Isolation forest based KNN graph}
\usage{
calculate_knn_graph(feature_mat, k, hardwareSetup)
}
\arguments{
\item{feature_mat}{A feature matrix based on which KNN graph is computed}

\item{k}{Number of neighbours in KNN graph.}

\item{hardwareSetup}{Hardware setup for parallelising isolation forest}
}
\value{
edgeList
}
\description{
Calculate a KNN graph from a feature matrix and
prune the graph using isolation forest.
}
