% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{calculate_knn_graph_distance_matrix_Stats}
\alias{calculate_knn_graph_distance_matrix_Stats}
\title{One standard-deviation based KNN graph (Single-thread vversion)}
\usage{
calculate_knn_graph_distance_matrix_Stats(feature_mat, k)
}
\arguments{
\item{feature_mat}{A feature matrix based on which KNN graph is computed}

\item{k}{Number of neighbours in KNN graph}
}
\value{
edgeList
}
\description{
Calculate a KNN graph from a feature matrix using stats one-std based methods.
}
