% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAE.R
\name{GCNModelVAE}
\alias{GCNModelVAE}
\title{GCN Model Variational Auto-enccoder auto-encoding variational Bayes}
\usage{
GCNModelVAE(input_feat_dim, hidden_dim1, hidden_dim2, dropout)
}
\arguments{
\item{input_feat_dim}{dimension of input feature matrix}

\item{hidden_dim1}{dimension of matrix in the first hidden layer}

\item{hidden_dim2}{dimension of matrix in the second hidden layer}

\item{dropout}{dropout}
}
\value{
A new \code{GCNModelAE} object.
}
\description{
Create an GCN auto-encoder object.
}
\details{
Auto-encoder for GCN using

#' @references
\insertRef{GCN}{scRGNet}
}
\section{Fields}{

\describe{
\item{\code{gc1}}{the first GCN layer}

\item{\code{gc2}}{the second GCN layer}

\item{\code{gc3}}{the third GCN layer}

\item{\code{dc}}{Decoder using inner product for prediction}
}}

