% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts}
\alias{gene_counts}
\title{scRNA-seq gene expression counts from GSE138852}
\format{
A matrix with columns:
\describe{
\item{AD}{Human Entorhinal Cortex of aged individuals with Alzheimer’s disease (AD)}
\item{Healthy}{Human Entorhinal Cortex of aged Healthy individuals, the control.}
}
}
\source{
Duke-NUS Medical School, Singapore.
}
\usage{
gene_counts
}
\description{
A single-cell atlas of the human cortex reveals drivers of transcriptional changes in Alzheimer’s disease in specific cell subpopulations
}
\examples{
\dontrun{
 gene_counts
}
}
\references{
\preformatted{Grubman A, Chew G, Ouyang JF, Sun G et al. A single-cell atlas of entorhinal cortex from individuals with Alzheimer's disease reveals cell-type-specific gene expression regulation. Nat Neurosci 2019 Dec;22(12):2087-2097. PMID: 31768052
}
}
\keyword{datasets}
