% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateGraph}
\alias{generateGraph}
\title{Generate a cell graph}
\usage{
generateGraph(feature_mat, k = 7)
}
\arguments{
\item{feature_mat}{A feature matrix}

\item{k}{Number of neighbours}
}
\description{
Compute a cell graph from a feature matrix. More methods will be added in future.
}
