% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAE.R
\name{GraphCovolution}
\alias{GraphCovolution}
\title{Graph Convolutional Neural Network Layer}
\usage{
GraphCovolution(in_features, out_features, dropout = 0, act = torch::nnf_relu)
}
\arguments{
\item{in_features}{in_features.}

\item{out_features}{out_features}

\item{dropout}{default 0}

\item{act}{The activation function used by the neuron. Default ReLU}
}
\value{
A new \code{GraphCovolution} object.

Loss
}
\description{
Create a new GCN layer object.

Forward feeding

Reset weight
}
\details{
A GCN layer class

#' @references
\insertRef{GCN}{scRGNet}
}
\section{Fields}{

\describe{
\item{\code{in_features}}{in_features}

\item{\code{out_features}}{out_features}

\item{\code{dropout}}{dropout}

\item{\code{act}}{activation function}

\item{\code{weight}}{weight of the model}
}}

