% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss_function_gcn}
\alias{loss_function_gcn}
\title{GCN Loss function}
\usage{
loss_function_gcn(
  VAE = FALSE,
  logvar = NULL,
  preds,
  labels,
  mu,
  n_nodes,
  norm,
  pos_weight
)
}
\arguments{
\item{VAE}{Whether to use variational auto-encoder or not. Default False.}

\item{logvar}{Used when VAE is set to TRUE. Default is NULL when using AE.}

\item{preds}{Predictions made by GCN during training}

\item{labels}{True values to checked against preds to calculate loss}

\item{mu}{Mean of the Gaussian for auto-encoder}

\item{n_nodes}{Number of nodes in the graph}

\item{norm}{Norm}

\item{pos_weight}{Weight parameter used to calculate loss}
}
\value{
Logistic-cross-entropy loss

#' @references
\insertRef{scGNN}{scRGNet}
}
\description{
Loss function for optimising GCN using either AE or VAE
}
