% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss_function_graph}
\alias{loss_function_graph}
\title{Loss function regularised by the graph information}
\usage{
loss_function_graph(
  recon_x,
  x,
  mu,
  logvar,
  graph_regu = NULL,
  gamma_param = 1,
  regu_mat = NULL,
  regu_type = NULL,
  regu_param = 0.001,
  model = "AE",
  reduction = "sum"
)
}
\arguments{
\item{recon_x}{Reconstructed scRNA-seq matrix}

\item{x}{Original input scRNA-seq matrix}

\item{mu}{mu}

\item{logvar}{logvar}

\item{graph_regu}{graph_regu}

\item{gamma_param}{gamma_param}

\item{regu_mat}{reg_mat}

\item{regu_type}{Type of regulariser}

\item{regu_param}{reg_param}

\item{model}{AE}

\item{reduction}{sum

#' @references
\insertRef{scGNN}{scRGNet}}
}
\description{
Reconstruction + KL divergence losses summed over all elements and batch
}
