% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellNet}
\alias{plotCellNet}
\title{Plot a cell network}
\usage{
plotCellNet(net, group = TRUE)
}
\arguments{
\item{net}{A graph object}

\item{group}{Whether to show cell community in the network}
}
\description{
plot
}
