% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDegree}
\alias{plotDegree}
\title{Plot a cell network}
\usage{
plotDegree(net, title = "Distribution of Vertices in the Cell Network")
}
\arguments{
\item{net}{A graph object}

\item{title}{Title of the histogram}
}
\description{
plot
}
