% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLogRank}
\alias{plotLogRank}
\title{Plot a cell network}
\usage{
plotLogRank(net, title = "A log-log Plot of Connectivities for Cell Network")
}
\arguments{
\item{net}{A graph object}

\item{title}{Title of the log-log plot}
}
\description{
plot
}
