% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{regulation_mse_loss_function}
\alias{regulation_mse_loss_function}
\title{Regulation MSE Loss Function}
\usage{
regulation_mse_loss_function(input, target, regu_mat, reduction = "none")
}
\arguments{
\item{input}{tensor}

\item{target}{tensor}

\item{regu_mat}{Matrix}

\item{reduction}{String}
}
\value{
tensor
}
\description{
Measures the element-wise mean squared error for regulation input, now only support LTMG.
}
