% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runFeatureAE}
\alias{runFeatureAE}
\title{Main function of running scGNN feature auto-encoder}
\usage{
runFeatureAE(
  scDataset,
  LTMG_mat = NULL,
  hyperParams = list(batch_size = 1L, regu_epochs = 5L, L1 = 0.5, L2 = 0.5, regu_alpha
    = 0.9, reduction = "sum"),
  hardwareSetup = list(CUDA = F, coresUage = 5L)
)
}
\arguments{
\item{scDataset}{A scDataset object containing scRNA-seq data}

\item{LTMG_mat}{LTMG sparse matrix. If provided, then LTMG regularisation will be applied. Optional.}

\item{hyperParams}{A list of hyperparameter to tune the model. Optional.}

\item{hardwareSetup}{A list of parameters to setup the hardware on which the model runs. Optional.}
}
\value{
A matrix representing the encoded space of the scRNA-seq matrix.
}
\description{
The main function for performing dimensional reduction on the preprocessed scRNA-seq data
}
\examples{
# Example 1:
# Tested examples. Not run for fast package compiling
\dontrun{
# Accessing the demo gene_counts_small dataset available with the package
inputCountsPath <- system.file("extdata", "GSE138852_small.csv", package = "scRGNet")
# Preprocess the raw counts
counts <- preprocessCSV(path = inputCountsPath)
ltmg <- runLTMG(counts)
hyperParams <- setHyperParams(regu_epochs = 5L)
hardwareSetup <- setHardware(coresUsage = 1L)
z <- runFeatureAE(scDataset = counts, LTMG_mat = ltmg, hyperParams, hardwareSetup)
}

}
\references{
\insertRef{scGNN}{scRGNet}
\insertRef{LTMG}{scRGNet}
\insertRef{torch}{scRGNet}
\insertRef{progress}{scRGNet}
\insertRef{matrix}{scRGNet}
}
