% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessCSV.R
\name{runLTMG}
\alias{runLTMG}
\title{Get discretised regulatory signals from LTMG model}
\usage{
runLTMG(
  expr_mat = NULL,
  fromFile = FALSE,
  readPath = NULL,
  toFile = TRUE,
  outdir_path = file.path(getwd(), "output"),
  saveName = "ltmg"
)
}
\arguments{
\item{expr_mat}{A data frame or a matrix}

\item{fromFile}{A Boolean value}

\item{readPath}{A string}

\item{toFile}{A Boolean value}

\item{outdir_path}{A string}

\item{saveName}{A string}
}
\value{
Returns an LTMG object
}
\description{
A wrapper function for running scGNNLTMG that uses
the Left-Trunctruncated-Mixed-Gaussian(LTMG) model to
translate the input gene expression data pre-processed by preprocessCSV into
a discretised regulatory signal as the regularizer for
the feature autoencoder of scGNN.
It will take around 10-15 minutes for inferring LTMG tags.
This step is optional but highly recommended prior to scGNN analysis.
}
\references{
\insertRef{LTMG}{scRGNet}
\insertRef{scGNN}{scRGNet}
}
