% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runSCGNN}
\alias{runSCGNN}
\title{Main function of running scGNN framework for analysis}
\usage{
runSCGNN(
  datasetName,
  outputDir,
  LTMGregularized = FALSE,
  LTMGpath = NULL,
  hyperParams = list(EM_iteration = 10, regu_epochs = 500, EM_epochs = 200, k = 7,
    GAEepochs = 200, L1 = 1, L2 = 0)
)
}
\arguments{
\item{datasetName}{A character vector representing the name of the dataset used}

\item{outputDir}{Path to the output directory}

\item{LTMGregularized}{Whether to use LTMG tags as regularizer}

\item{LTMGpath}{Path to the LTMG sparse mtx file}

\item{hyperParams}{A list of hyperparameter to tune the model. Optional.}
}
\description{
Main
}
