% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscRGNet.R
\name{runscRGNet}
\alias{runscRGNet}
\title{Launch Shiny webtool}
\usage{
runscRGNet()
}
\value{
No return value but open up a Shiny page.
}
\description{
Launch Shiny webtool
}
\examples{
\dontrun{
scRGNet::runscRGNet()
}

}
\references{
\insertRef{shiny}{scRGNet}
}
