% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scDataset}
\alias{scDataset}
\title{scRNA-seq dataset class}
\usage{
scDataset(data, transform = NULL)
}
\arguments{
\item{data}{A matrix of pre-processed scRNA-seq data, with genes as rows and samples as columns}

\item{transform}{A function to transform the matrix}
}
\value{
A new \code{scDataset} object.

number of cells

A list of sample information
\itemize{
\item index  - Index of sample in the feature matrix
\item sample - Name of the sample
\item expr   - Gene expression values in the sample
}
}
\description{
Create a new scDataset object.

Get the number of cells in the dataset

Get the expression values of a cell
}
\details{
An R6 class to store scRNA-seq data for analysis using Torch
}
\references{
\insertRef{scGNN}{scRGNet}
\insertRef{torch}{scRGNet}
\insertRef{r6}{scRGNet}
}
