% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{setHardware}
\alias{setHardware}
\title{A hardware setter to run the model}
\usage{
setHardware(coresUsage = 1L, CUDA = FALSE)
}
\arguments{
\item{coresUsage}{The number of CPU cores you would like to use to train the model.
For Mac user, please do not change the default value of 1. This is a special case for Mac OS (sorry)}

\item{CUDA}{Whther to use Nvidia GPU to train the model via CUDA.
Make sure you have CUDA 10.0 or 11.1 version installed.
Higher version of CUDA is currently not supported by torch.
If CUDA is set to TRUE, coresUsage will not be used. Default FALSE.}
}
\value{
A list of parameters to set up hardware
}
\description{
Setting up your hardware to train the model faster and get a better result (if you have a powerful one)
}
\references{
\insertRef{scGNN}{scRGNet}
}
