% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{setHyperParams}
\alias{setHyperParams}
\title{Hyper-parameter setter}
\usage{
setHyperParams(
  batch_size = 1L,
  regu_epochs = 5L,
  L1 = 0.5,
  L2 = 0.5,
  regu_alpha = 0.9,
  reduction = "sum"
)
}
\arguments{
\item{batch_size}{Batch size used to train the auto-encoder}

\item{regu_epochs}{Epoch of trainning; that is, the number of time the dataset is visited.}

\item{L1}{Intensity of L1 penalty on loss if LTMG matrix is used. Takes value in range >=0 and >= 1}

\item{L2}{Intensity of L2 penalty on loss if LTMG matrix is used. Takes value in range >=0 and >= 1}

\item{regu_alpha}{Intensity of LTMG regularisation. Takes value in range >=0 and >= 1}

\item{reduction}{Type of reduction used in loss functions:
available options: "mean", "sum", or "none", meaning not using reduction method.}
}
\value{
A list of hyper-parameters for runFeatureAE
}
\description{
A function to set up the hyper-parameter to tune the auto-encoder
}
\references{
\insertRef{scGNN}{scRGNet}
}
