% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{vallina_mse_loss_function}
\alias{vallina_mse_loss_function}
\title{Vallina MSE}
\usage{
vallina_mse_loss_function(input, target, reduction = "none")
}
\arguments{
\item{input}{tensor}

\item{target}{tensor}

\item{reduction}{Reduction method}
}
\value{
tensor
}
\description{
Calculate vallina mse loss
}
