% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts_small}
\alias{gene_counts_small}
\title{A subset scRNA-seq data of GSE138852}
\format{
A matrix with columns:
\describe{
\item{AD}{Human Entorhinal Cortex of aged individuals with Alzheimer’s disease (AD)}
\item{Healthy}{Human Entorhinal Cortex of aged Healthy individuals, the control.}
}
}
\source{
Duke-NUS Medical School, Singapore.
}
\usage{
gene_counts_small
}
\description{
A subset of scRNA-seq gene counts from data set GSE138852. It is a small subset of the data with only 500 genes for running a quick demo of data processing functions of the package.
}
\examples{
\dontrun{
 gene_counts_small
}
}
\references{
\preformatted{Grubman A, Chew G, Ouyang JF, Sun G et al. A single-cell atlas of entorhinal cortex from individuals with Alzheimer's disease reveals cell-type-specific gene expression regulation. Nat Neurosci 2019 Dec;22(12):2087-2097. PMID: 31768052
}
}
\keyword{datasets}
