% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateNetwork}
\alias{generateNetwork}
\title{Generate a cell network}
\usage{
generateNetwork(feature_mat, k = 7)
}
\arguments{
\item{feature_mat}{A feature matrix}

\item{k}{Number of neighbours}
}
\description{
Compute a cell network from a feature matrix. More methods will be added in future.
}
