% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateNetwork}
\alias{generateNetwork}
\title{Infer a cell-cell network from the encoded space using one-stat KNN}
\usage{
generateNetwork(feature_mat, k = NULL, hardwareSetup = NULL)
}
\arguments{
\item{feature_mat}{An encoded matrix generated by the feature auto-encoder.}

\item{k}{Number of neighbours in KNN graph. If k is not provided,
then the best heuristic k = floor(sqrt(N)) will be used,
where N is the sample size.}

\item{hardwareSetup}{Hardware setup for running isolation forest}
}
\value{
An igraph object to plot the cell network.
}
\description{
Compute a cell network from a feature matrix using KNN.
}
\examples{
# Example 1:
# Tested examples. Not run for fast package compiling
\dontrun{
# Accessing the demo gene_counts_small dataset available with the package
inputCountsPath <- system.file("extdata", "GSE138852_small.csv", package = "scRGNet")
# Preprocess the raw counts
counts <- preprocessCSV(path = inputCountsPath)
ltmg <- runLTMG(counts)
hyperParams <- setHyperParams(regu_epochs = 5L)
hardwareSetup <- setHardware(coresUsage = 1L)
z <- runFeatureAE(scDataset = counts, LTMG_mat = ltmg, hyperParams, hardwareSetup)
net <- generateNetwork(z)
}

}
\references{
\insertRef{scGNN}{scRGNet}
\insertRef{isotree}{scRGNet}
\insertRef{isolationForest}{scRGNet}
\insertRef{stats}{scRGNet}
\insertRef{defaultK}{scRGNet}
\insertRef{igraph}{scRGNet}
}
