% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runFeatureAE}
\alias{runFeatureAE}
\title{Main function of running scGNN feature auto-encoder for analysis}
\usage{
runFeatureAE(
  scDataset,
  LTMG_mat = NULL,
  hyperParams = list(batch_size = 1L, regu_epochs = 5L, L1 = 0.5, L2 = 0.5, regu_alpha
    = 0.9, reduction = "sum"),
  hardwareSetup = list(CUDA = F, coresUage = 5L)
)
}
\arguments{
\item{scDataset}{A scDataset object containing scRNA-seq data}

\item{LTMG_mat}{LTMG sparse matrix. If provided, then LTMG regularisation will be applied. Optional.}

\item{hyperParams}{A list of hyperparameter to tune the model. Optional.}

\item{hardwareSetup}{A list of parameters to setup the hardware on which the model runs. Optional.}
}
\description{
Main
}
