% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runSCGNN}
\alias{runSCGNN}
\title{Main function of running scGNN framework for analysis}
\usage{
runSCGNN(
  dataset,
  outputDir,
  LTMG_mat = NULL,
  hyperParams = list(batch_size = 12800L, EM_iteration = 10L, regu_epochs = 500L,
    EM_epochs = 200L, K = 7L, GAEepochs = 200L, L1 = 1, L2 = 0),
  hardwareSetup = list(CUDA = FALSE, coresUage = 5)
)
}
\arguments{
\item{dataset}{A scDataset object containing scRNA-seq data}

\item{outputDir}{Path to the output directory}

\item{LTMG_mat}{LTMG sparse matrix. If provided, then LTMG regularisation will be applied. Optional.}

\item{hyperParams}{A list of hyperparameter to tune the model. Optional.}

\item{hardwareSetup}{A list of parameters to setup the hardware on which the model runs. Optional.}
}
\description{
Main
}
