% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scDataset}
\alias{scDataset}
\title{scRNA-seq dataset class}
\usage{
scDataset(data, transform = NULL)
}
\arguments{
\item{data}{A matrix of pre-processed scRNA-seq data, with genes as rows and samples as columns}

\item{transform}{A function to transform the matrix}
}
\value{
A new \code{scDataset} object.

number of cells
}
\description{
Create a new scDataset object.

Get the number of cells in the dataset

Get the expression values of a cell
}
\details{
An R6 class to store scRNA-seq data for analysis using Torch
}
\section{Fields}{

\describe{
\item{\code{features}}{A feature matrix with rows as cells and column as genes}

\item{\code{Any}}{function to transform the matrix. Optional.}
}}

