% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Main Training Model}
\usage{
train(
  epoch,
  train_loader,
  model,
  optimiser,
  regu_mat = NULL,
  regu_param = 0.9,
  reduction = "sum",
  device,
  EMflag = FALSE,
  EMreguTag = FALSE
)
}
\arguments{
\item{epoch}{epoch}

\item{train_loader}{A torch dataloader loaded with an scDataset object}

\item{model}{A feature auto-encoder object}

\item{optimiser}{Optimiser}

\item{regu_mat}{LTMG regulation matrix}

\item{regu_param}{Regularisation paramter alpha}

\item{reduction}{reduction type: mean/sum, default(sum)}

\item{device}{hardware to train the model}

\item{EMflag}{Emflag}

\item{EMreguTag}{whether to use regu_mat in EM process

#' @references
\insertRef{scGNN}{scRGNet}}
}
\description{
Train in main
}
