% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Main Training Model}
\usage{
train(
  epoch,
  train_loader,
  model,
  optimiser,
  regu_mat = NULL,
  hyperParams,
  device,
  EMflag = FALSE
)
}
\arguments{
\item{epoch}{epoch}

\item{train_loader}{A torch dataloader loaded with an scDataset object}

\item{model}{A feature auto-encoder object}

\item{optimiser}{Optimiser}

\item{regu_mat}{LTMG regulation matrix}

\item{hyperParams}{Hyper-parameters user defined to tune the model}

\item{device}{hardware to train the model}

\item{EMflag}{Emflag

#' @references
\insertRef{scGNN}{scRGNet}}
}
\description{
Train in main
}
