% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Training Model}
\usage{
train(epoch, train_loader, model, optimiser, device, EMflag = FALSE)
}
\arguments{
\item{epoch}{epoch}

\item{train_loader}{A torch dataloader loaded with an scDataset object}

\item{model}{A feature auto-encoder object}

\item{optimiser}{Optimiser}

\item{device}{hardware to train the model}

\item{EMflag}{Emflag

#' @references
\insertRef{scGNN}{scRGNet}}
}
\description{
Train in main
}
